Move the products actions to the store branch
